Move existing tests from overview to show file
and couple of tests for collapsed view of description
