+++ test/components/project/overview/description.spec.js
@@ -1,13 +1,15 @@
 import { RouterLinkStub } from '@vue/test-utils';
 
+import { nextTick } from 'vue';
 import ProjectOverviewDescription from '../../../../src/components/project/overview/description.vue';
 
 import { mockRouter } from '../../../util/router';
 import { mount } from '../../../util/lifecycle';
 
-const mountComponent = (props) => mount(ProjectOverviewDescription, {
+const mountComponent = (props, attachTo) => mount(ProjectOverviewDescription, {
   props,
-  container: { router: mockRouter('/projects/1') }
+  container: { router: mockRouter('/projects/1') },
+  attachTo
 });
 
 describe('ProjectOverviewDescription', () => {
@@ -39,5 +41,22 @@ describe('ProjectOverviewDescription', () => {
   it('renders router link to project settings', () => {
     const { to } = mountComponent({ canUpdate: true }).getComponent(RouterLinkStub).props();
     to.should.equal('/projects/1/settings');
+  });
+
+  describe('collapsed', () => {
+    const description = 'line1\nline2\nline3\nline4\nline5';
+
+    it('is collapsed if there are more than 4 lines of description', async () => {
+      const component = mountComponent({ description, canUpdate: false }, document.body);
+      await nextTick();
+      component.find('#project-overview-description').attributes('style').should.match(/height: .*px/);
+    });
+
+    it('should expand on click', async () => {
+      const component = mountComponent({ description, canUpdate: false }, document.body);
+      await nextTick();
+      await component.find('#project-overview-description').trigger('click');
+      component.find('#project-overview-description').attributes('style').should.match(/height: auto/);
+    });
   });
 });
