+++ test/components/project/show.spec.js
@@ -1,3 +1,4 @@
+import ProjectOverviewDescription from '../../../src/components/project/overview/description.vue';
 import Loading from '../../../src/components/loading.vue';
 import NotFound from '../../../src/components/not-found.vue';
 import ProjectOverview from '../../../src/components/project/overview.vue';
@@ -130,6 +131,44 @@ describe('ProjectShow', () => {
       await loadLocale(app.vm.$container, 'es');
       await app.vm.$nextTick();
       title.text().should.equal('My Project (archivado)');
+    });
+  });
+
+  describe('project description', () => {
+    it('allows admins to see instructions about editing', async () => {
+      mockLogin({ role: 'admin' });
+      testData.extendedProjects.createPast(1);
+      const app = await load('/projects/1');
+      const desc = app.getComponent(ProjectOverviewDescription);
+      desc.props().description.should.equal('');
+      desc.props().canUpdate.should.equal(true);
+    });
+
+    it('allows managers to see instructions about editing', async () => {
+      mockLogin({ role: 'none' });
+      testData.extendedProjects.createPast(1, { role: 'manager' });
+      const app = await load('/projects/1');
+      const desc = app.getComponent(ProjectOverviewDescription);
+      desc.props().description.should.equal('');
+      desc.props().canUpdate.should.equal(true);
+    });
+
+    it('does not allow viewers to see instructions about editing', async () => {
+      mockLogin({ role: 'none' });
+      testData.extendedProjects.createPast(1, { role: 'viewer' });
+      const app = await load('/projects/1', {}, { deletedForms: false });
+      const desc = app.getComponent(ProjectOverviewDescription);
+      desc.props().description.should.equal('');
+      desc.props().canUpdate.should.equal(false);
+    });
+
+    it('passes project description through', async () => {
+      mockLogin({ role: 'admin' });
+      testData.extendedProjects.createPast(1, { description: 'Description' });
+      const app = await load('/projects/1');
+      const desc = app.getComponent(ProjectOverviewDescription);
+      desc.props().description.should.equal('Description');
+      desc.props().canUpdate.should.equal(true);
     });
   });
